Download Helm revision manifests and values

+++ pkg/app/helm/values.go
@@ -6,7 +6,7 @@ import (
 	"github.com/angelokurtis/kts-cli/pkg/bash"
 )
 
-func GetValues(release string, revision int, options ...OptionFunc) ([]byte, error) {
+func GetValues(release string, revision int64, options ...OptionFunc) ([]byte, error) {
 	o := new(Option)
 	if err := o.apply(options...); err != nil {
 		return nil, err
